Split CatalogLoadError into its own module; map the middleware-refactor reuse
One module, one class, per repo style: CatalogLoadError moves out of
hocon_catalog_cache.py, with imports updated in the cache and both
consuming middlewares (no test imports the error directly).

Also documents, in the cache's module docstring, how the planned
recasting of GetToolbox/GetSubnetwork/GetMcpTool as designer middlewares
relates to this class: same SharedProcessCache substrate and the
MiddlewareInfoMiddleware injection template are reusable, but each tool
deliberately differs in a policy this class fixes (special-purpose
restorers, empty-result semantics, freshness), and their network fan-out
halves are not file catalogs at all. Each tool carries a pointer to that
map so it cannot drift in four places.
